refactor: Move InventoryClient out into clients/ while keeping bridge adapters in clients/adapters/
