+++ src/clients/bridges/BaseAdapter.ts
