+++ test/InventoryClient.InventoryRebalance.ts
@@ -3,7 +3,8 @@ import { toBN, toWei, randomAddress, spyLogIncludes } from "./utils";
 
 import { InventoryConfig } from "../src/interfaces";
 import { MockBundleDataClient, MockHubPoolClient, MockAdapterManager, MockTokenClient } from "./mocks/";
-import { CrossChainTransferClient, InventoryClient } from "../src/clients"; // Tested
+import { InventoryClient } from "../src/clients"; // Tested
+import { CrossChainTransferClient } from "../src/clients/bridges";
 
 const toMegaWei = (num: string | number | BigNumber) => ethers.utils.parseUnits(num.toString(), 6);
